+++ com.unity.cinemachine/Runtime/Core/CameraBlendStack.cs
@@ -63,7 +63,7 @@ namespace Unity.Cinemachine
         {
             public int Id;
             public int Priority;
-            public readonly CinemachineBlend Source = new (null, null, null, 0, 0);
+            public readonly CinemachineBlend Source = new ();
             public float DeltaTimeOverride;
 
             // If blending in from a snapshot, this holds the source state
@@ -71,7 +71,7 @@ namespace Unity.Cinemachine
             ICinemachineCamera m_SnapshotSource;
             float m_SnapshotBlendWeight;
 
-            public StackFrame() : base(new (null, null, null, 0, 0)) {}
+            public StackFrame() : base(new ()) {}
             public bool Active => Source.IsValid;
 
             // This is a little tricky, because we only want to take a new snapshot at the start 
@@ -121,18 +121,20 @@ namespace Unity.Cinemachine
 
             var frame = m_FrameStack[FindFrame(overrideId, priority)];
             frame.DeltaTimeOverride = deltaTime;
-            frame.Source.CamA = camA;
-            frame.Source.CamB = camB;
-            frame.Source.Duration = 1;
-            frame.Source.BlendCurve = s_DefaultLinearAnimationCurve;
             frame.Source.TimeInBlend = weightB;
 
-            // In case vcams are inactive game objects, make sure they get initialized properly
-            if (camA is CinemachineVirtualCameraBase vcamA)
-                vcamA.EnsureStarted();
-            if (camB is CinemachineVirtualCameraBase vcamB)
-                vcamB.EnsureStarted();
+            if (frame.Source.CamA != camA || frame.Source.CamB != camB)
+            {
+                frame.Source.CustomBlender = CinemachineCore.GetCustomBlender?.Invoke(camA, camB);
+                frame.Source.CamA = camA;
+                frame.Source.CamB = camB;
 
+                // In case vcams are inactive game objects, make sure they get initialized properly
+                if (camA is CinemachineVirtualCameraBase vcamA)
+                    vcamA.EnsureStarted();
+                if (camB is CinemachineVirtualCameraBase vcamB)
+                    vcamB.EnsureStarted();
+            }
             return overrideId;
             
             // local function to get the frame index corresponding to the ID
@@ -148,7 +150,10 @@ namespace Unity.Cinemachine
                 for (index = 1; index < count; ++index)
                     if (m_FrameStack[index].Priority > priority)
                         break;
-                m_FrameStack.Insert(index, new StackFrame { Id = withId, Priority = priority });
+                var frame = new StackFrame { Id = withId, Priority = priority };
+                frame.Source.Duration = 1;
+                frame.Source.BlendCurve = s_DefaultLinearAnimationCurve;
+                m_FrameStack.Insert(index, frame);
                 return index;
             }
         }
@@ -197,18 +202,22 @@ namespace Unity.Cinemachine
             else
             {
                 var frame = m_FrameStack[0];
-                frame.Blend.CamA = frame.Blend.CamB = null;
-                frame.Source.CamA = frame.Source.CamB = null;
+                frame.Blend.ClearBlend();
+                frame.Blend.CamB = null;
+                frame.Source.ClearBlend();
+                frame.Source.CamB = null;
             }
         }
 
         /// <summary>
         /// Call this every frame with the current active camera of the root frame.
         /// </summary>
+        /// <param name="context">The mixer context in which this blend stack exists</param>
         /// <param name="activeCamera">Current active camera (pre-override)</param>
         /// <param name="up">Current world up</param>
         /// <param name="deltaTime">How much time has elapsed, for computing blends</param>
         public void UpdateRootFrame(
+            ICinemachineMixer context,
             ICinemachineCamera activeCamera, Vector3 up, float deltaTime)
         {
             // Make sure there is a first stack frame
@@ -221,6 +230,7 @@ namespace Unity.Cinemachine
             if (activeCamera != outgoingCamera)
             {
                 bool backingOutOfBlend = false;
+                float duration = 0;
 
                 // Do we need to create a game-play blend?
                 if (LookupBlendDelegate != null && activeCamera != null && activeCamera.IsValid
@@ -235,18 +245,24 @@ namespace Unity.Cinemachine
 
                         frame.Source.CamA = outgoingCamera;
                         frame.Source.BlendCurve = blendDef.BlendCurve;
+                        duration = blendDef.BlendTime;
                     }
-                    frame.Source.Duration = blendDef.BlendTime;
+                    frame.Source.Duration = duration;
                     frame.Source.TimeInBlend = 0;
+                    frame.Source.CustomBlender = null;
                 }
                 frame.Source.CamB = activeCamera;
+
+                // Get custom blender, if any
+                if (duration > 0)
+                    frame.Source.CustomBlender = CinemachineCore.GetCustomBlender?.Invoke(outgoingCamera, activeCamera);
 
                 // Update the working blend:
                 // Check the status of the working blend.  If not blending, no problem.
                 // Otherwise, we need to do some work to chain the blends.
-                var duration = frame.Source.Duration;
+                ICinemachineCamera camA;
                 if (frame.Blend.IsComplete)
-                    frame.Blend.CamA = frame.GetSnapshotIfAppropriate(outgoingCamera, 0); // new blend
+                    camA = frame.GetSnapshotIfAppropriate(outgoingCamera, 0); // new blend
                 else
                 {
                     bool snapshot = (frame.Blend.State.BlendHint & CameraState.BlendHints.FreezeWhenBlendingOut) != 0;
@@ -258,7 +274,10 @@ namespace Unity.Cinemachine
                         && frame.Blend.Uses(activeCamera))
                         snapshot = true;
 
+                    // Avoid nesting too deeply
                     var nbs = frame.Blend.CamA as NestedBlendSource;
+                    if (!snapshot && nbs != null && nbs.Blend.CamA is NestedBlendSource nbs2)
+                        nbs2.Blend.CamA = new SnapshotBlendSource(nbs2.Blend.CamA);
 
                     // Special case: if backing out of a blend-in-progress
                     // with the same blend in reverse, adjust the blend time
@@ -273,42 +292,55 @@ namespace Unity.Cinemachine
                             duration += sbs.RemainingTimeInBlend;
                     }
 
-                    // Avoid nesting too deeply
-                    if (!snapshot && nbs != null && nbs.Blend.CamA is NestedBlendSource nbs2)
-                        nbs2.Blend.CamA = new SnapshotBlendSource(nbs2.Blend.CamA);
-
                     // Chain to existing blend
-                    frame.Blend.CamA = snapshot 
-                        ? new SnapshotBlendSource(frame, frame.Blend.Duration - frame.Blend.TimeInBlend)
-                        : new NestedBlendSource(new CinemachineBlend(frame.Blend));
+                    if (snapshot)
+                        camA = new SnapshotBlendSource(frame, frame.Blend.Duration - frame.Blend.TimeInBlend);
+                    else
+                    {
+                        var blendCopy = new CinemachineBlend();
+                        blendCopy.CopyFrom(frame.Blend);
+                        camA = new NestedBlendSource(blendCopy);
+                    }
                 }
+                // For the event, we use the raw outgoing camera, not the blend source
+                frame.Blend.CamA = outgoingCamera;
                 frame.Blend.CamB = activeCamera;
                 frame.Blend.BlendCurve = frame.Source.BlendCurve;
                 frame.Blend.Duration = duration;
                 frame.Blend.TimeInBlend = 0;
+                frame.Blend.CustomBlender = frame.Source.CustomBlender;
+
+                // Allow the client to modify the blend
+                if (duration > 0)
+                    CinemachineCore.BlendCreatedEvent.Invoke(new () { Origin = context, Blend = frame.Blend });
+
+                // In case the event handler tried to change the cameras, put them back
+                frame.Blend.CamA = camA;
+                frame.Blend.CamB = activeCamera;
             }
 
             // Advance the working blend
-            AdvanceBlend(frame.Blend, deltaTime);
+            if (AdvanceBlend(frame.Blend, deltaTime))
+                frame.Source.ClearBlend();
             frame.UpdateCameraState(up, deltaTime);
 
             // local function
-            static void AdvanceBlend(CinemachineBlend blend, float deltaTime)
+            static bool AdvanceBlend(CinemachineBlend blend, float deltaTime)
             {
+                bool blendCompleted = false;
                 if (blend.CamA != null)
                 {
                     blend.TimeInBlend += (deltaTime >= 0) ? deltaTime : blend.Duration;
                     if (blend.IsComplete)
                     {
                         // No more blend
-                        blend.CamA = null;
-                        blend.BlendCurve = null;
-                        blend.Duration = 0;
-                        blend.TimeInBlend = 0;
+                        blend.ClearBlend();
+                        blendCompleted = true;
                     }
                     else if (blend.CamA is NestedBlendSource bs)
                         AdvanceBlend(bs.Blend, deltaTime);
                 }
+                return blendCompleted;
             }
         }
 
@@ -343,9 +375,19 @@ namespace Unity.Cinemachine
                     // Handle cases where we're blending into or out of nothing
                     if (!frame.Blend.IsComplete)
                     {
-                        frame.Blend.CamA ??= frame.GetSnapshotIfAppropriate(
-                            m_FrameStack[lastActive], frame.Blend.TimeInBlend);
-                        frame.Blend.CamB ??= m_FrameStack[lastActive];
+                        if (frame.Blend.CamA == null)
+                        {
+                            frame.Blend.CamA = frame.GetSnapshotIfAppropriate(
+                                m_FrameStack[lastActive], frame.Blend.TimeInBlend);
+                            frame.Blend.CustomBlender = CinemachineCore.GetCustomBlender?.Invoke(
+                                frame.Blend.CamA, frame.Blend.CamB);
+                        }
+                        if (frame.Blend.CamB == null)
+                        {
+                            frame.Blend.CamB = m_FrameStack[lastActive];
+                            frame.Blend.CustomBlender = CinemachineCore.GetCustomBlender?.Invoke(
+                                frame.Blend.CamA, frame.Blend.CamB);
+                        }
                     }
                     lastActive = i;
                 }
@@ -355,7 +397,8 @@ namespace Unity.Cinemachine
 
         /// <summary>
         /// Set the current root blend.  Can be used to modify the root blend state.
-        /// <param name="blend">The new blend.  If null, current blend will be force-completed.</param>
+        /// <param name="blend">The new blend.  CamA and CamB will be ignored, but the other fields 
+        /// will be taken.  If null, current blend will be force-completed.</param>
         /// </summary>
         public void SetRootBlend(CinemachineBlend blend)
         {
@@ -364,7 +407,12 @@ namespace Unity.Cinemachine
                 if (blend == null)
                     m_FrameStack[0].Blend.Duration = 0;
                 else
-                    m_FrameStack[0].Blend.CopyFrom(blend);
+                {
+                    m_FrameStack[0].Blend.BlendCurve = blend.BlendCurve;
+                    m_FrameStack[0].Blend.TimeInBlend = blend.TimeInBlend;
+                    m_FrameStack[0].Blend.Duration = blend.Duration;
+                    m_FrameStack[0].Blend.CustomBlender = blend.CustomBlender;
+                }
             }
         }